+++ components/x-increment/stories/knobs.js
@@ -1,0 +1,7 @@
+module.exports = (data, { number }) => {
+	return {
+		count() {
+			return number('Count', data.count, {});
+		}
+	};
+};