+++ components/x-interaction/src/Interaction.jsx
@@ -1,17 +1,18 @@
 import { h } from '@financial-times/x-engine';
 import { InteractionClass } from './InteractionClass';
-import { InteractionRender } from './InteractionRender';
+import { InteractionSSR } from './InteractionSSR';
 import wrapComponentName from './concerns/wrap-component-name';
 import { registerComponent } from './concerns/register-component';
 
 // use the class version for clientside and the static version for server
-const Interaction = typeof window !== 'undefined' ? InteractionClass : InteractionRender;
+const Interaction = typeof window !== 'undefined' ? InteractionClass : InteractionSSR;
 
 export const withActions = (getActions) => (Component) => {
 	function Enhanced({
 		id,
-		hydrating,
 		actions: extraActions,
+		actionsRef,
+		serialiser,
 		...initialState
 	}) {
 		// support passing actions to withActions as an object or a function
@@ -24,7 +25,8 @@ export const withActions = (getActions) => (Component) => {
 			id,
 			Component,
 			initialState,
-			hydrating,
+			actionsRef,
+			serialiser,
 			// if extraActions is defined, those are from another level
 			// of wrapping with withActions, so those should take precedence
 			actions: Object.assign(actions, extraActions),
@@ -40,5 +42,6 @@ export const withActions = (getActions) => (Component) => {
 	return Enhanced;
 };
 
-export { default as hydrate } from './Hydrate';
-export { getInteractionSerialiser } from './concerns/serialiser';
+export { hydrate } from './Hydrate';
+export { HydrationData } from './HydrationData';
+export { Serialiser } from './concerns/serialiser';