+++ components/x-interaction/src/InteractionClass.jsx
@@ -1,6 +1,5 @@
 import { h, Component } from '@financial-times/x-engine';
 import { InteractionRender } from './InteractionRender';
-import { loading } from './concerns/symbols';
 import mapValues from './concerns/map-values';
 
 export class InteractionClass extends Component {
@@ -8,7 +7,8 @@ export class InteractionClass extends Component {
 		super(props, ...args);
 
 		this.state = {
-			[loading]: false,
+			state: {},
+			inFlight: 0,
 		};
 
 		this.actions = mapValues(props.actions, (func) => (...args) => {
@@ -16,28 +16,43 @@ export class InteractionClass extends Component {
 			// setting loading back to false will happen in the same microtask and no
 			// additional render will be scheduled.
 			Promise.resolve().then(() => {
-				this.setState({ [loading]: true });
+				this.setState(({ inFlight }) => ({ inFlight: inFlight + 1 }));
 			});
 
-			Promise.resolve(func(...args)).then((next) => {
-				if(typeof next === 'function') {
-					this.setState(
-						state => next(Object.assign(
+			return Promise.resolve(func(...args)).then((next) => {
+				const updater = typeof next === 'function'
+					? ({state}) => ({state: Object.assign(
+						state,
+						next(Object.assign(
 							{},
-							this.props.initialState,
-							state,
+							props.initialState,
+							state
 						))
-					);
-				} else {
-					this.setState(next);
-				}
+					)})
+					: ({state}) => ({state: Object.assign(state, next)});
 
-				this.setState({ [loading]: false });
+				return new Promise(resolve =>
+					this.setState(updater, () => (
+						this.setState(({ inFlight }) => ({ inFlight: inFlight - 1 }), resolve)
+					))
+				);
 			});
 		});
 	}
 
+	componentDidMount() {
+		if(this.props.actionsRef) {
+			this.props.actionsRef(this.actions);
+		}
+	}
+
+	componentWillUnmount() {
+		if(this.props.actionsRef) {
+			this.props.actionsRef(null);
+		}
+	}
+
 	render() {
-		return <InteractionRender {...this.props} state={this.state} actions={this.actions} />;
+		return <InteractionRender {...this.props} {...this.state} actions={this.actions} />;
 	}
 }