+++ components/x-teaser/Props.d.ts
@@ -9,23 +9,23 @@ export type Theme = 'extra-article';
 
 export type Modifier = 'stacked' | 'centre' | 'stretched' | 'opinion-background' | 'landscape' | 'big-story' | string;
 
-export type ImageSize = 'XS' | 'Small' | 'Medium' | 'Large' | 'XL';
+export type ImageSize = 'XS' | 'Small' | 'Medium' | 'Large' | 'XL' | 'XXL';
 
 export interface Features {
-	showMeta: boolean;
-	showTitle: boolean;
-	showStandfirst: boolean;
-	showStatus: boolean;
-	showImage: boolean;
-	showHeadshot: boolean;
-	showVideo: boolean;
-	showRelatedLinks: boolean;
-	showCustomSlot: boolean;
+	showMeta?: boolean;
+	showTitle?: boolean;
+	showStandfirst?: boolean;
+	showStatus?: boolean;
+	showImage?: boolean;
+	showHeadshot?: boolean;
+	showVideo?: boolean;
+	showRelatedLinks?: boolean;
+	showCustomSlot?: boolean;
 }
 
 export interface General {
 	id: string;
-	url: string;
+	url?: string;
 	/** Preferred to url if available */
 	relativeUrl?: string;
 	type: ContentType;
@@ -69,12 +69,12 @@ export interface Image {
 	/** Images must be accessible to the Origami Image Service */
 	image?: Media;
 	imageSize?: ImageSize;
-	imageLazyload?: Boolean;
+	imageLazyload?: Boolean | String;
 }
 
 export interface Headshot {
-	headshot?: Media;
-	headshotTint?: 'string'
+	headshot?: String;
+	headshotTint?: String;
 }
 
 export interface Video {