+++ components/x-teaser/stories/article.js
@@ -69,7 +69,6 @@ exports.knobs = [
 	'headlineTesting',
 	// Variants
 	'layout',
-	'theme',
 	'modifiers',
 ];