+++ components/x-teaser/stories/index.js
@@ -1,13 +1,14 @@
 const { Teaser } = require('../');
 
 exports.component = Teaser;
+
 exports.package = require('../package.json');
 
 exports.dependencies = {
-	'o-date': '^v2.11.0',
-	'o-fonts': '^3.0.0',
+	'o-normalise': '^1.6.0',
+	'o-date': '^2.11.0',
 	'o-typography': '^5.5.0',
-	'o-teaser': '^2.2.5',
+	'o-teaser': '^2.3.0',
 	'o-labels': '^3.0.0',
 	'o-video': '^4.1.0',
 };
@@ -16,6 +17,7 @@ exports.stories = [
 	require('./article'),
 	require('./opinion'),
 	require('./package'),
+	require('./package-item'),
 	require('./promoted'),
 	require('./top-story'),
 	require('./video'),