+++ packages/x-babel-config/jest.js
@@ -1,0 +1,4 @@
+const getBabelConfig = require('./');
+const babelJest = require('babel-jest');
+
+module.exports = babelJest.createTransformer(getBabelConfig());