+++ tools/x-docs/gatsby-config.js
@@ -1,50 +1,55 @@
-const path = require('path');
-
 module.exports = {
+	// The GitHub Pages deployment will be in this sub-folder
 	pathPrefix: '/x-dash',
+	polyfill: false,
 	siteMetadata: {
-		title: 'Gatsby Default Starter',
+		title: 'x-dash',
+		siteUrl: 'https://financial-times.github.io/x-dash/',
+		description: 'Shared front-end for FT.com and The App.'
 	},
 	plugins: [
-		'gatsby-plugin-react-next',
-		'gatsby-plugin-react-helmet',
-		'gatsby-plugin-packages',
 		{
-			resolve: 'gatsby-plugin-sass',
+			resolve: 'gatsby-source-filesystem',
 			options: {
-				includePaths: path.resolve(__dirname, 'bower_components')
-			}
+				name: 'docs',
+				path: './src/data'
+			},
 		},
-		'gatsby-plugin-munge-pages',
+		{
+			resolve: 'gatsby-source-filesystem',
+			options: {
+				name: 'docs',
+				path: '../../docs'
+			},
+		},
+		{
+			resolve: 'gatsby-source-filesystem',
+			options: {
+				name: 'components',
+				path: '../../components'
+			},
+		},
+		{
+			resolve: 'gatsby-source-filesystem',
+			options: {
+				name: 'packages',
+				path: '../../packages'
+			},
+		},
+		// Handles markdown files (creates "MarkdownRemark" nodes)
 		{
 			resolve: 'gatsby-transformer-remark',
 			options: {
 				plugins: [
 					'gatsby-remark-prismjs',
 					'gatsby-remark-autolink-headers',
-					'gatsby-remark-external-links',
-					'gatsby-remark-emoji',
-					'gatsby-remark-file-links',
-				],
-			},
+					'gatsby-remark-external-links'
+				]
+			}
 		},
-		'gatsby-plugin-doc-pages',
-		// {
-		// 	resolve: '@andrew-codes/gatsby-plugin-elasticlunr-search',
-		// 	options: {
-		// 		fields: [
-		// 			'title',
-		// 			'plainText',
-		// 		],
-		// 		resolvers: {
-		// 			MarkdownRemark: {
-		// 				title: node => node.frontmatter.title,
-		// 				plainText: node => node.plainText,
-		// 				href: node => node.frontmatter.path,
-		// 				breadcrumbs: node => node.frontmatter.breadcrumbs,
-		// 			}
-		// 		}
-		// 	}
-		// }
+		// Handles package manifest files (creates "NpmPackage" nodes)
+		'gatsby-transformer-npm-package',
+		// Handles YAML files (creates "YourFileNameYaml" nodes)
+		'gatsby-transformer-yaml'
 	]
 };