+++ changelog.md
@@ -1,5 +1,25 @@
 # Changelog
 
-## v1.0.0-beta.1
+## v1
+
+### v1.0.0-beta.3
+
+- Removes whitespace output around text output in x-teaser (#146)
+- Refactors headshot prop to accept a string value in x-teaser (#153)
+- Refactors theme indicator and theme rule to enable explicit overrides in x-teaser (#154)
+- Adds parentTheme prop to x-teaser (#154)
+
+
+### v1.0.0-beta.2
+
+- Adds support for fragments to x-engine (#134)
+- Removes obsolete syndication props from x-teaser (#137)
+- Adds support for configurable image lazy loading to x-teaser (#135)
+- Refactors rendering of title prop to avoid stringification in x-teaser (#139)
+- Updates presets to enable relative time by default in x-teaser (#142)
+- Adds XXL image size prop to x-teaser (#143)
+- Refactors teaser rulesets to avoid minification issues in x-teaser (#144)
+
+### v1.0.0-beta.1
 
 - Initial public release of x-engine, x-interaction, x-handlebars, and x-teaser